tamil_panchangam_engine/requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anthropic==0.86.0
anyio==4.12.1
attrs==25.4.0
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.6
click==8.3.1
distro==1.9.0
docstring_parser==0.17.0
duckdb==1.5.0
psycopg2-binary==2.9.9
fastapi==0.135.1
flatbuffers==25.12.19
h11==0.16.0
h3==4.4.2
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
idna==3.11
jiter==0.13.0
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
numpy==2.4.3
pillow==12.1.1
pycparser==3.0
pydantic==2.12.5
pydantic_core==2.41.5
pyswisseph==2.10.3.2
python-dotenv==1.2.2
pytz==2026.1.post1
PyYAML==6.0.3
referencing==0.37.0
regex==2026.2.28
reportlab==4.4.10
requests==2.32.5
rpds-py==0.30.0
sniffio==1.3.1
SQLAlchemy==2.0.48
sqlmodel==0.0.37
slowapi==0.1.9
limits==3.7.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
google-auth==2.29.0
starlette==0.52.1
svglib==1.4.1
svgwrite==1.4.3
swisseph==0.0.0.dev1
tiktoken==0.12.0
timezonefinder==8.2.1
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
uvicorn==0.42.0
uvloop==0.22.1
watchfiles==1.1.1
websockets==16.0
